class_path/c10_kafta_db/microservice_01/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Aqeel Shahzad <67902032+aqeel-spec@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
fastapi = "^0.111.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlmodel = "^0.0.18"
psycopg = {extras = ["binary"], version = "^3.1.19"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"




